Integrate new binAST parser into lazy parsing/compilation pipeline
This diff integrates the new parser into the background compiler worker thread pool
mechanism, enabling us to run it in the background and pass the changes back via
the SharedFunctionInfo structure. We do this via a similar method as the PreparserData.
This diff also add the initial skeleton of the serialization and deserialization
of the binary AST, thus enabling us to properly pass an Isolate-independent AST that
deserializes directly into the standard internal AST, allowing us to completely
skip re-parsing the JS source on the main thread. Most of the serializer/deserializer
have yet to be implemented.