Add serialization support for AstConsStrings
This requires serializing a series of references to the component AstRawStrings.
Since they can be variable length, we needed to refactor the return type of the
deserialization functions so they can return both the deserialized value along
with the new offset. Also includes a new test that we can run inside d8 to test
serialization/deserialization.