Add support for proper deserialization of Scope locals_
Previously we weren't using the correct Declare method to add to the locals_
variable and we were serializing in the order of the VariableMap instead of
in the order of locals_. We were also doing some stuff incorrectly that was
triggering an assert when serializing AstRawStrings for named functions.
This diff fixes those issues and modifies the test file to trigger it in the future.