Add de/serialization support for ExpressionStatement and VariableProxyExpression
ExpressionStatement is pretty straightforward. VariableProxyExpression (and
VariableProxies along with them) required some reworking of how we handle Variables.
Originally we were storing a nested map of Variables and their IDs by Scope.
This actually served little purpose during the process of de/serialization, so this
diff removes that first layer, so now all Variable/ID pairs are stored in a single
pool during de/serialization. This allows us to deserialize Variable definitions that
aren't tied to any particular Scope (i.e. var->scope() == nullptr), which was required
for VariableProxy objects that referenced resolved Variables. If there was a non-null
Scope in the Variable, then that Scope's Variable declarations should already have been
processed and its Variables should already live in the pool, thus only requiring a reference.
However, if the Variable did not belong to any known Scope (as is often the case with
eval-ed scripts), we would need to define a new Variable in place which was not bound
to any Scope.

Another aspect of this diff is rebuilding the chain of unresolved VariableProxies.
We needed to be able to encode a way to reference the other VariableProxy objects.
To make things simple we use the position of the VariableProxy in the source as a
unique way to refer to it. Because we may not have de/serialized all VariableProxies
by the time we encounter them via the unresolved chain, we needed to separate the
linking of the unresolved chain from the deserialization of the individual VariableProxies.