Compress serialized AST with zlib
Since we store the serialized AST for an arbitrary amount of time before
using it during lazy parsing/compilation, it's important to reduce its
memory footprint as much as we can while keeping deserialization on the
main thread fast. Using zlib, we can compress the serialized AST on the
background parsing thread while maintaining good performance on the main
thread when we need to decompress and deserialize.

This diff also changes the internal interface for the deserializer to use
a raw uint8_t array rather than the ByteArray object. This was necessary
since zlib operates on raw arrays of bytes, and we'll eventually move away
from ByteArray entirely in the future to save memory.