Gracefully handle unimplemented AST nodes in the serializer
Previously we would either silently continue or crash if we encountered
an AST node that we didn't support. Now we note that we encountered an
unimplemented node and simply discard the results. This allows us to
still get serializer coverage while also allowing us to handle scripts
that contain features we haven't added support for yet.