Add ability to easily turn compression on/off
After some simple performance experiments we realized that compression was
a pretty big performance regression. There are still some experiments we want
to perform, but for now this diff makes it easy to enable/disable and disables
it by default. Also some other performance related changes (the new inlines
file for the deserializer) as well as miscellaneous cleanup.