Add de/serializer support for ForInStatement nodes
This completes our binAST support for ES5 loops. Relatively straightforward,
but required fixing a bug related to how we were serializing Variables.
Namely it was possible for there to be Variables contained in the Scope's
locals_ that weren't in the main variables_ map (due to some desugaring logic
that created Temporary Variables), but the code implicitly assumed otherwise.