Add varint encoding for string and variable indices
The encoding for these uint32 indices is done in 7 bit chunks with an
the extra top bit indicating whether there is another byte that follows.
On our test script this reduced the size of our uncompressed serialized
AST by varying amounts, from ~10% up to 33%. If I had to guess, a typical
function would probably be closer to 10% savings.