Fix several bugs in AST de/serialization logic
This diff fixes a number of bugs encountered when trying to run the
speculative parser on real websites.

1) We had inverted the condition for excluding ASTs with DeclarationScopes
that had rare data objects. This led to exluding most of the functions we
*could* have serialized, and only serializing those which we actually didn't
support.
2) There was an off-by-one error in reading the flag fields for DeclarationScopes
which led to us incorrectly treating loading from "this", which caused downstream
mismatches in the generated bytecode, among other things.
3) Expression has a bit field (is_parenthesized) that isn't set by default when
creating new AstNodes that are subtypes of Expression. This would lead to bit_field
mismatches, which we had added DCHECKs for. We now instead explicitly overwrite
the bit_field with whatever we deserialized to ensure that they match and to prevent
this kind of issue in the future.
4) There was a typo where we were passing the source position as the offset when
deserializing boolean literals, causing downstream misinterpretation due to a newly
misaligned offset into the serialized byte array.