Fix deserialization of FunctionDeclarations to properly handle scope chain
We were deserializing the FunctionLiterals inside FunctionDeclarations as
part of the deserialization of the DeclarationScope. The scope being currently
deserialized needed to be in the scope chain of the resulting FunctionLiteral,
but it wasn't the active scope. This diff adds a usage of FunctionState to
push the scope onto the scope stack so it is properly referenced by the FunctionLiteral.

Also fixes a random GC crash due to not handling BinAstParseData and friend
inside a switch statement for object descriptors.