Fix a couple issues with temporary variable handling
We were using an unordered_set to store names for locals and temporaries (which are a subset of locals).
We would then use the total size of the sets to determine how many non-local variables to serialize.
This obviously wouldn't work in the presence of variable name collisions (as was the case in functions with
multiple "invisible" temporaries with empty names). We fixed this by explicitly counting how many locals
and temporaries there in addition to using unordered sets for fast lookups.

There was another issue on the deserialization side in that we were adding the temporary locals to the
set of scope variables (rather than just the scope's locals), which was causing us to generate incorrect bytecode.
We fixed this in the deserializer by distinguishing between temporary and non-temporary locals when
deserializing the scope's variable map.