Add support for BreakStatement & ContinueStatement
BreakStatement and ContinueStatement reference other BreakableStatement nodes within
the AST. We'd like to represent these references as offsets into the serialized buffer,
however these referenced nodes could come before or after the Break/ContinueStatements
in the AST, i.e. we may not have an offset for them yet. This diff introduces some
additional bookkeeping done by both the serializer and the deserializer to track
the offsets of BreakableStatements as well as the offsets of things we need to patch
because we didn't have an offset available to us when we first processed the item.
When we do eventually process that BreakableStatement, we go back to placeholders
and patch them with the now-available offset/address.