Add proxy string table for each FunctionLiteral
Previously every function that we deserialized had to first deserialize
the string table at the beginning of the serialized AST. Since we serialize
a single monolithic buffer for a functions within a top-level nested function
tree, this led to functions having to deserialize lots of strings that they
would never use.

After this diff, while serializing we track which strings are used by each
function literal. We then record the offsets of those strings at the beginning
of the portion of the serialized buffer for that function literal. This way,
when we deserialize a sub-section of the buffer corresponding to an inner function
we only need to deserialize the strings at those offsets in the global string
table table.