Fix de/serialization for ObjectLiteral and TryCatchStatement
These two bugs were encountered while browsing around the web.
We were trying to re-compute how many boilerplate properties a particular
ObjectLiteral had upon deserialization instead of just storing the field
in the serialization format. Doing the latter fixed the crash. We also
were not behaving correctly in the absence of a Scope when serializing
a TryCatchStatement, which we now mark correctly.