Use full width reads and writes during de/serialization
During the initial phase of development of the preparse++ prototype,
we chose to implement reading fields larger than a byte in a naive
fashion (reading a single byte at a time). This diff changes things
to use the proper type of the read/write so that we take advantage
of the full data width the processor support.