Make string metadata fixed size to speed up deserialization
Prior to this diff the strings in the global string table were variable
size due to storing their contents inline. To avoid having to deserialize
all raw strings within a function family, we maintained a proxy string table
for each function that served as a forwarding pointer to the offset in the
global string table for each string used within that function. During
deserialization we needed to construct this map for lookups, which ended
up being expensive in deserialization microbenchmarks.

To make these lookups cheaper and to avoid having to allocate this hash map,
this diff stores the string header and contents separately. This enables the
function-local proxy strings to store an index into the global string table
rather than a global offset. Now during deserialization we construct a small
vector of raw strings for just the strings referenced within this function
and index into it.