Add function-local proxy table for Variables
To avoid having to pay the cost of several hash lookups in the global
Variable table during deserialization of Variables within a scope,
this diff adds a per-function table with pointers to the global
variable table. This enables us to create a dense vector of
Variables during deserialization, eliminating the aforementioned
hash lookups.