Unhack inner function skipping
Previously we added a few utility functions to help us determine when we
should try to skip functions during deserialization using the ConsumedPreparseData.
This was an unreliable method, causing us to crash due to some edge cases within
the serialized PreparseData format. Ideally we would have some way to tell that
a function was skipped originally during preparsing so we wouldn't have to guess
from the contents of the PreparseData. Fortunately, this already existed in the
form of the is_skipped_function field on the DeclarationScope which we were
intentionally ignoring previously! Wow, lucky us. So now that we use that field
we can remove a bunch of hacky code from the previous implementation.