Limit dispatching background parse tasks to top-level parsing
We were always dispatching background parse tasks, even when performing
lazy parsing on the main thread, which was significantly impacting parsing
performance when we were unable to use the deserializer. This was in turn
making overall time spent parsing on the main thread much worse, even if
deserialization performance was better. With the current design of the
background speculative parser, we only need to enqueue a task when parsing
the outermost function lazy function. By limiting our enqueuing to when
the original scope was a script scope (i.e. a top-level scope), we avoid
trying to enqueue a bunch of redundant background tasks, thus reducing
the work the main thread and background threads need to do.